// src/api/types/database.types.ts
export interface Database {
  public: {
    Tables: {
      users: {
        Row: {
// ...
          posts: string[];
// ...
        };
// ...
      };
// ...
      posts: {
        Row: {
          id: string;
          type: string;
          category: string;
          author: string;
          content: Json;
          visibility: string;
          status: string;
          tags: string[];
          game_tags: string[];
          mentions: string[];
          likes: string[];
          comments: string[];
          shares: number;
          saves: number;
          reports: string[];
// ...
        };
// ...
      };
// ...
    };
// ...
  };
}
// ...
export type Post = Database['public']['Tables']['posts']['Row'];
export type PostInsert = Database['public']['Tables']['posts']['Insert'];
export type PostUpdate = Database['public']['Tables']['posts']['Update'];